converted fixed4 to half4

+++ Assets/Scenes/BenjaminSwee_Tutorial/Shaders/NewUnlitShader.shader
@@ -40,7 +40,7 @@ Shader "Custom/URP_BaseShader"
                 return o;
             }
 
-            fixed4 frag (v2f i) : SV_Target
+            half4 frag (v2f i) : SV_Target
             {
                 // sample the texture
                 fixed4 col = tex2D(_MainTex, i.uv);
